Server class modifications
